use ROR autocomplete for research facilities
