+++ app/views/stash_datacite/contributors/_facility_form.html.erb
@@ -6,7 +6,7 @@
 %>
 
 <div class="c-input">
-  <%= react_component('components/MetadataEntry/FacilityAutocomplete',
+  <%= react_component('components/MetadataEntry/FacilityForm',
                       {name: (contributor.contributor_name || ''),
                        rorId: (contributor.name_identifier_id || ''),
                        contribId: contributor.id,
